refactor: extract app wizard and workflow helpers
